Refac

+++ pacman-ui-fx-2d/src/main/java/de/amr/games/pacman/ui/fx/rendering2d/PacManGameGhostAnimations.java
@@ -101,7 +101,7 @@ public class PacManGameGhostAnimations extends SpriteAnimations {
 	public void select(String name, Object... args) {
 		super.select(name, args);
 		if (Ghost.ANIM_GHOST_NUMBER.equals(name)) {
-			animation(Ghost.ANIM_GHOST_NUMBER).setFrameIndex((Integer) args[0]);
+			animation(Ghost.ANIM_GHOST_NUMBER).setFrameIndex((Byte) args[0]);
 		}
 	}
